Consolidate hash functions
